\name{extract.annotation.mach}
\alias{extract.annotation.mach}
\title{extracts SNP annotation from MACH/HapMap legend files...}
\usage{extract.annotation.mach(infofile, legendfile, chromosome=NA)}
\description{extracts SNP annotation from MACH/HapMap legend files}
\details{This function extracts SNP annotation from MACH info and HapMap legend files.}
\value{data frame containing annotaton}
\author{Yurii Aulchenko}
\keyword{IO}
\keyword{manip}
\arguments{\item{infofile}{MACH (ml)info-file name}
\item{legendfile}{HapMap legend file name}
\item{chromosome}{chromosome}}
