\name{ge03d2c}
\alias{ge03d2c}
\docType{data}
\title{GWA-type data on small region}
\description{
	\code{ge03d2c} 
	A small data set (approximately 200 people and 8,000 SNPs) containing 
	data on 3 autosomes and X chromsome. This data set is 
	complementary to \code{\link{ge03d2}}.
}
\usage{data(ge03d2c)}
\format{
}
\details{
}
\source{
}
\references{
}
\examples{
#main example: use this to see full functionality
# demo(ge03d2)

# load and work with ge03d2c
data(ge03d2c)
mc <- check.marker(ge03d2c)
summary(mc)
}
\keyword{datasets}
