\name{ge03d2ex}
\alias{ge03d2ex}
\docType{data}
\title{GWA-type data on small region}
\description{
	\code{ge03d2ex} 
	A small data set (approximately 150 people and 4,000 SNPs) containing 
	data on 3 autosomes and X chromsome. Is a good set for 
	demonatration of the QC procedures (different genotyping errors
	are introduced) and GWA analysis. 
	This data set was developed for the "Advances in population-
	based studies" (Ge03) course of the Nihes.
	See vignette "GenABEL-tutorial.pdf" for details.
}
\usage{data(ge03d2ex)}
\format{
}
\details{
}
\source{
}
\references{
}
\examples{
#main example: use this to see full functionality
# demo(ge03d2ex)

# load and work with ge03d2ex
data(ge03d2exex)
}
\keyword{datasets}
