\name{snps.cell-class}
\docType{class}
\alias{snps.cell-class}

\title{Class "snps.cell"}
\description{
This is a lowest-level class based on which 
\code{\link{snp.mx-class}} is build
}
\author{Yurii Aulchenko}
\note{User is not supposed to work with this class. Use \code{\link{snp.data-class}}.}
\seealso{
	\code{\link{snp.mx-class}},
	\code{\link{gwaa.data-class}},
	\code{\link{snp.data-class}}
}
%\examples{}
\keyword{classes}

