\name{summary.scan.gwaa}
\alias{summary.scan.gwaa}
\title{Shortcut to 'descriptives...}
\usage{
\method{summary}{scan.gwaa}(object, ...) 
}
\description{Shortcut to 'descriptives.scan'}
\details{Shortcut to 'descriptives.scan'}
\author{Yurii Aulchenko}
\arguments{\item{object}{object of class 'scan.gwaa' as generated by 
\code{\link{qtscore}}, \code{\link{mlreg}}, etc.}
\item{...}{arguments passed to \code{\link{descriptives.scan}}}}
\examples{
require(GenABEL.data)
data(srdta)
x <- qtscore(qt2,srdta)
summary(x)}
